Update Authenticated User Profile  Reset Account Password with the email

+++ src/main/java/org/soroth/procuctapi/dto/request/UpdateProductRequest.java
@@ -1,8 +1,10 @@
 package org.soroth.procuctapi.dto.request;
 
+import java.math.BigDecimal;
+
 public record UpdateProductRequest(
         String name,
         String description,
-        Float price
+        BigDecimal price
 ) {
 }
